Project reorganization. - Split app into /todo /chat and /index folders - Separated todo and global (user, api) context - Added @ alias for resolving imports (../../colors.ts -> @/colors.ts isnt it nice?) - Added react-router-dom, and defined a few routes - Added Index page which will contain all apps lavender-snake team has
