+++ client/package.json
@@ -22,7 +22,11 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
+    "react-router-dom": "^6.9.0",
     "typescript": "^4.9.3",
     "vite": "^4.1.0"
+  },
+  "devDependencies": {
+    "@types/node": "^18.15.3"
   }
 }
