+++ client/src/Context.tsx
@@ -1,11 +1,6 @@
 import { GoogleOAuthProvider } from "@react-oauth/google";
-import { createContext, useContext, useEffect, useState } from "react";
-import { useApiClient } from "./apiClient";
-import { Todo } from "./types";
+import { createContext, useContext, useState } from "react";
 import { isProd } from "./utils";
-
-const TodoContext = createContext<[Todo[] | undefined, any]>([[], () => {}]);
-export const useTodos = () => useContext(TodoContext);
 
 export const ApiEndpoints = {
 	GCP: "https://lavender-snake-server-wgikiljsnq-uc.a.run.app",
@@ -47,23 +42,6 @@ const UserContext = createContext<[User | undefined, any]>([
 ]);
 export const useUser = () => useContext(UserContext);
 
-function TodoContextProvider({ children }: any) {
-	const [todos, setTodos] = useState<Todo[] | undefined>(undefined);
-	const apiClient = useApiClient();
-	useEffect(() => {
-		if (!apiClient.user) return;
-		(async () => {
-			const t = await apiClient.getTodos();
-			setTodos(t);
-		})();
-	}, [apiClient]);
-
-	return (
-		<TodoContext.Provider value={[todos, setTodos]}>
-			{children}
-		</TodoContext.Provider>
-	);
-}
 function UserContextProvider({ children }: any) {
 	const userState = useState<User | undefined>(undefined);
 	const [apiEndpoint] = useApiEndpoint();
@@ -87,7 +65,7 @@ export function Context({ children }: any) {
 		<GoogleOAuthProvider clientId="987357728284-k763p5d1paujmlvn83tl2lc637f8ofc3.apps.googleusercontent.com">
 			<ApiEndpointContext.Provider value={[apiEndpoint, setApiEndpoint]}>
 				<UserContextProvider>
-					<TodoContextProvider>{children}</TodoContextProvider>
+					{children}
 				</UserContextProvider>
 			</ApiEndpointContext.Provider>
 		</GoogleOAuthProvider>
