+++ client/src/index/IndexContent.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const IndexContent = () => {
+	return (
+		<div>
+			<a href={"/todo"}>todo</a>
+		</div>
+	);
+};
+
+export default IndexContent;
